move Netgen statushandler into ngcore
