+++ libsrc/core/ngcore.hpp
@@ -23,5 +23,10 @@
 #include "ngstream.hpp"
 #include "utils.hpp"
 #include "ranges.hpp"
+#include "statushandler.hpp"
 
 #endif // NETGEN_CORE_NGCORE_HPP
+
+
+
+
