+++ libsrc/meshing/global.cpp
@@ -52,7 +52,6 @@ namespace netgen
   int silentflag = 0;
   int testmode = 0;
 
-  volatile multithreadt multithread;
 
   string ngdir = ".";
 
@@ -85,18 +84,6 @@ namespace netgen
 
   int h_argc = 0;
   char ** h_argv = NULL;
-
-  multithreadt :: multithreadt()
-  {
-    pause =0;
-    testmode = 0;
-    redraw = 0;
-    drawing = 0;
-    terminate = 0;
-    running = 0;
-    percent = 0;
-    task = "";
-  }
 
   DebugParameters debugparam;
   bool verbose = 0;
