+++ libsrc/meshing/global.hpp
@@ -31,23 +31,6 @@ namespace netgen
 
   DLL_HEADER extern mutex tcl_todo_mutex;
 
-  class DLL_HEADER multithreadt
-  {
-  public:
-    int pause;
-    int testmode;
-    int redraw;
-    int drawing;
-    int terminate;
-    int running;
-    double percent;
-    const char * task;
-    bool demorunning;
-    string * tcl_todo = new string("");  // tcl commands set from parallel thread
-    multithreadt();
-  };
-
-  DLL_HEADER extern volatile multithreadt multithread;
 
   class DebugParameters;
   class Mesh;
